+++ packages/core/src/Mark.ts
@@ -137,15 +137,6 @@ declare module '@tiptap/core' {
     }) => void) | null,
 
     /**
-     * The view has changed.
-     */
-    onViewUpdate?: ((this: {
-      options: Options,
-      editor: Editor,
-      type: MarkType,
-    }) => void) | null,
-
-    /**
      * The editor state has changed.
      */
     onTransaction?: ((