+++ packages/core/src/extensions/commands.ts
@@ -24,6 +24,7 @@ import * as liftListItem from '../commands/liftListItem'
 import * as newlineInCode from '../commands/newlineInCode'
 import * as replace from '../commands/replace'
 import * as replaceRange from '../commands/replaceRange'
+import * as resetAttributes from '../commands/resetAttributes'
 import * as resetNodeAttributes from '../commands/resetNodeAttributes'
 import * as scrollIntoView from '../commands/scrollIntoView'
 import * as selectAll from '../commands/selectAll'
@@ -43,6 +44,7 @@ import * as toggleWrap from '../commands/toggleWrap'
 import * as undoInputRule from '../commands/undoInputRule'
 import * as unsetAllMarks from '../commands/unsetAllMarks'
 import * as unsetMark from '../commands/unsetMark'
+import * as updateAttributes from '../commands/updateAttributes'
 import * as updateNodeAttributes from '../commands/updateNodeAttributes'
 import * as wrapIn from '../commands/wrapIn'
 import * as wrapInList from '../commands/wrapInList'
@@ -72,6 +74,7 @@ export { liftListItem }
 export { newlineInCode }
 export { replace }
 export { replaceRange }
+export { resetAttributes }
 export { resetNodeAttributes }
 export { scrollIntoView }
 export { selectAll }
@@ -91,6 +94,7 @@ export { toggleWrap }
 export { undoInputRule }
 export { unsetAllMarks }
 export { unsetMark }
+export { updateAttributes }
 export { updateNodeAttributes }
 export { wrapIn }
 export { wrapInList }
@@ -125,6 +129,7 @@ export const Commands = Extension.create({
       ...newlineInCode,
       ...replace,
       ...replaceRange,
+      ...resetAttributes,
       ...resetNodeAttributes,
       ...scrollIntoView,
       ...selectAll,
@@ -144,6 +149,7 @@ export const Commands = Extension.create({
       ...undoInputRule,
       ...unsetAllMarks,
       ...unsetMark,
+      ...updateAttributes,
       ...updateNodeAttributes,
       ...wrapIn,
       ...wrapInList,