+++ packages/react/src/NodeViewContent.tsx
@@ -1,5 +1,4 @@
 import React from 'react'
-import { useReactNodeView } from './useReactNodeView'
 
 export interface NodeViewContentProps {
   className?: string,
@@ -7,14 +6,12 @@ export interface NodeViewContentProps {
 }
 
 export const NodeViewContent: React.FC<NodeViewContentProps> = props => {
-  const { isEditable } = useReactNodeView()
   const Tag = props.as || 'div'
 
   return (
     <Tag
       className={props.className}
       data-node-view-content=""
-      contentEditable={isEditable}
       style={{ whiteSpace: 'pre-wrap' }}
     />
   )